feat: moving deeper into componentization, adjusted the result global state of resolve dice
