+++ src/context/CharProvider.js
@@ -3,6 +3,7 @@ import PropTypes from 'prop-types';
 import CharContext from './CharContext';
 
 const CharProvider = ({ children }) => {
+  const [hideSelection, setHideSelection] = useState(true);
   const [globalValue, setGlobalValue] = useState(0);
   const [enemyLife1, setEnemyLife1] = useState(8);
   const [enemyLife2, setEnemyLife2] = useState(5);
@@ -13,7 +14,9 @@ const CharProvider = ({ children }) => {
     enemyLife1,
     setEnemyLife1,
     enemyLife2,
-    setEnemyLife2
+    setEnemyLife2,
+    hideSelection,
+    setHideSelection,
   }
